Added new file to handle change events from dos
